refactor: Move 'AttendanceCard' to 'Card'
